+++ R/timeRange.R
@@ -5,10 +5,17 @@
 #' @param starttime Desired start datetime (ISO 8601).
 #' @param endtime Desired end datetime (ISO 8601).
 #' @param timezone Olson timezone used to interpret dates.
+#' @param unit Units used to determine time at end-of-day.
+#' @param ceilingEnd Logical instruction to apply
+#'   \code{\link[lubridate]{ceiling_date}} to the \code{enddate} rather than
+#'   \code{\link[lubridate]{floor_date}}
 #'
 #' @description
 #' Uses incoming parameters to return a pair of \code{POSIXct} times in the
-#' proper order.
+#' proper order. Both start and end times will have \code{lubridate::floor_date()}
+#' applied to get the nearest \code{unit} unless \code{ceilingEnd = TRUE} in
+#' which case the end time will will have \code{lubridate::ceiling_date()}
+#' applied.
 #'
 #' The required \code{timezone} parameter must be one of those found in
 #' \code{\link[base]{OlsonNames}}.
@@ -31,50 +38,41 @@
 timeRange <- function(
   starttime = NULL,
   endtime = NULL,
-  timezone = NULL
+  timezone = NULL,
+  unit = "sec",
+  ceilingEnd = FALSE
 ) {
 
-  # Validate parameters --------------------------------------------------------
+  # ----- Validate parameters --------------------------------------------------
 
-  if ( is.null(starttime) )
-    stop("Required parameter 'starttime' is missing.")
-
-  if ( is.null(endtime) )
-    stop("Required parameter 'endtime' is missing.")
-
-  if ( is.null(timezone) )
-    stop("Required parameter 'timezone' is missing.")
+  stopIfNull(starttime)
+  stopIfNull(endtime)
+  stopIfNull(timezone)
+  stopIfNull(unit)
+  stopIfNull(ceilingEnd)
 
   if ( !timezone %in% base::OlsonNames() )
     stop(paste0("Timezone '", timezone, "' is not recognized."))
 
-  # * Prepare POSIXct inputs ---------------------------------------------------
+  # ----- Process datetimes ----------------------------------------------------
 
-  ## NOTE on hadling POSIXct inputs:
-  #  When given a POSIXct time `lubridate::parse_date_time()` forces the time
-  #  into the timezone given to `lubridate::parse_date_time()`. This alters the
-  #  physical instant in time the original POSIXct represents, so we must
-  #  properly convert a POSIXct start or end date to the proper timezone before
-  #  passing it to `lubridate::parse_date_time()`
+  # Guarantee conversion to POSIXct
+  starttime <- parseDatetime(starttime, timezone = timezone)
+  endtime <- parseDatetime(endtime, timezone = timezone)
 
-  if ( lubridate::is.POSIXct(starttime) )
-    starttime <- lubridate::with_tz(starttime, tzone = timezone)
+  # Guarantee proper ordering
+  timeRange <- sort(c(starttime, endtime))
 
-  if ( lubridate::is.POSIXct(endtime) )
-    endtime <- lubridate::with_tz(endtime, tzone = timezone)
+  # Floor/Ceiling to nearest unit
+  timeRange[1] <- lubridate::floor_date(timeRange[1], unit = unit)
+  if ( ceilingEnd ) {
+    timeRange[2] <- lubridate::ceiling_date(timeRange[2], unit = unit)
+  } else {
+    timeRange[2] <- lubridate::floor_date(timeRange[2], unit = unit)
+  }
 
+  # ----- Return ---------------------------------------------------------------
 
-  # * Parse inputs -------------------------------------------------------------
-
-  orders <- c("Ymd", "YmdH", "YmdHM", "YmdHMS")
-
-  timeInputs <-
-    c(starttime, endtime) %>%
-    lubridate::parse_date_time(orders = orders, tz = timezone)
-
-
-  # Order output time limits ---------------------------------------------------
-
-  return(sort(timeInputs))
+  return(timeRange)
 
 }