+++ tests/testthat/test-initializeLogging.R
@@ -1,6 +1,6 @@
 test_that("initializeLogging() stops when logDir is invalid", {
 
-  expect_error(initializeLogging(), "Required parameter 'logDir' is missing.")
+  expect_error(initializeLogging(), "argument 'logDir' must not be NULL.")
 
 })